requirements.txt
pydantic
fastapi
pydantic
uvicorn
requests
gtts
firebase-admin
stripe
python-dotenv
elevenlabs

